+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 ## v4.1.0
 - Initial release of PipeDream PowerShell module v4
 - Like Star Wars Episodes I, II and III we dont talk about v1, v2 and v3
+
+## v4.2.0
+### :sparkles: New features
+- Added Dataverse HTTP functions:
+  - `Invoke-DataverseGet`
+  - `Invoke-DataversePost`
+  - `Invoke-DataversePatch`
+  - `Invoke-DataverseDelete`
+- Added ability to get the Dataverse URL from the bearer token